seeding, took out env to make the seeds work. Will need to add back once deploying
